[🎨] Refactor Spanish Translations implementation.
