--- src/main/java/infinituum/labellingcontainers/providers/language/spanish/SpanishArLangProvider.java
@@ -1,13 +1,0 @@
-package infinituum.labellingcontainers.providers.language.spanish;
-
-import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricLanguageProvider;
-
-public class SpanishArLangProvider extends FabricLanguageProvider {
-    public SpanishArLangProvider(FabricDataOutput dataOutput) { super(dataOutput, "es_ar"); }
-
-    @Override
-    public void generateTranslations(TranslationBuilder translationBuilder) {
-        new SpanishLangProvider(dataOutput).generateTranslations(translationBuilder);
-    }
-}
